% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-baselines.R
\name{.pairs_to_labels}
\alias{.pairs_to_labels}
\title{Convert a data.frame of pair-level match predictions to cluster labels}
\usage{
.pairs_to_labels(preds, id_vec, threshold = 0.5)
}
\description{
Convert a data.frame of pair-level match predictions to cluster labels
}
\keyword{internal}
