% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-baselines.R
\name{.python_script}
\alias{.python_script}
\title{Locate the inst/python directory for the package}
\usage{
.python_script(name)
}
\description{
Locate the inst/python directory for the package
}
\keyword{internal}
