% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-similarity.R
\name{.sim_jw}
\alias{.sim_jw}
\title{Jaro-Winkler similarity between two strings}
\usage{
.sim_jw(a, b)
}
\arguments{
\item{a, b}{Character scalars.}
}
\value{
Numeric in \eqn{[0,1]}, or \code{NA_real_}.
}
\description{
Jaro-Winkler similarity between two strings
}
\keyword{internal}
