% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{er_cosine_dist}
\alias{er_cosine_dist}
\title{Cosine distance on rows of a matrix}
\usage{
er_cosine_dist(X)
}
\arguments{
\item{X}{numeric matrix}
}
\value{
full cosine distance matrix with diag 0
}
\description{
Cosine distance on rows of a matrix
}
