% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-reporting.R
\name{er_draw_table}
\alias{er_draw_table}
\title{Draw a small table on a graphics device}
\usage{
er_draw_table(df, title = NULL, base_size = 8, max_rows_text = 60)
}
\description{
Draw a small table on a graphics device
}
