% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-eval-truth.R
\name{er_eval_ca_one}
\alias{er_eval_ca_one}
\title{Evaluate one set of labels against truth using GCMER metrics}
\usage{
er_eval_ca_one(
  pred,
  truth,
  metrics = c("chi2", "rand", "adj_rand", "fowlkes_mallow", "mirkin", "jaccard", "tpr",
    "fpr", "F_measure", "meila_heckerman", "max_match", "van_dongen", "mutual_info")
)
}
\arguments{
\item{pred}{integer vector}

\item{truth}{integer vector}

\item{metrics}{character vector of metric names}
}
\value{
tibble row with metrics
}
\description{
Evaluate one set of labels against truth using GCMER metrics
}
