% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_external_metrics}
\alias{er_external_metrics}
\title{External metrics (ARI, pairwise F1)}
\usage{
er_external_metrics(labels, truth_pairs = NULL, truth_vec = NULL)
}
\description{
External metrics (ARI, pairwise F1)
}
