% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-features.R
\name{er_features_tfidf_svd}
\alias{er_features_tfidf_svd}
\title{TF-IDF + truncated SVD features}
\usage{
er_features_tfidf_svd(text_vec, svd_dim = 100)
}
\arguments{
\item{text_vec}{character vector}

\item{svd_dim}{target dimension}
}
\value{
dense numeric matrix
}
\description{
TF-IDF + truncated SVD features
}
