% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-eval-truth.R
\name{er_get_clustering_agreement_fun}
\alias{er_get_clustering_agreement_fun}
\title{Locate clustering_agreement() (GCMER or sourced)}
\usage{
er_get_clustering_agreement_fun()
}
\value{
function
}
\description{
Locate clustering_agreement() (GCMER or sourced)
}
\keyword{internal}
