% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_internal_metrics}
\alias{er_internal_metrics}
\title{Internal metrics (Silhouette/CH/DB/Modularity/Stability)}
\usage{
er_internal_metrics(Z, labels, G = NULL, sample_n = 2000L)
}
\description{
Internal metrics (Silhouette/CH/DB/Modularity/Stability)
}
