% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_knn_graph}
\alias{er_knn_graph}
\title{Build cosine kNN graph with threshold}
\usage{
er_knn_graph(Z, k_knn = 50, min_sim = 0)
}
\arguments{
\item{Z}{matrix n x d}

\item{k_knn}{integer}

\item{min_sim}{numeric between 0 and 1}
}
\value{
igraph with weights = cosine similarity
}
\description{
Build cosine kNN graph with threshold
}
