% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-io.R
\name{er_load_input}
\alias{er_load_input}
\title{Load input by path, Excel, or data.frame}
\usage{
er_load_input(data, sheet = NULL)
}
\arguments{
\item{data}{data.frame or path or special key 'cora'}

\item{sheet}{sheet index/name for Excel}
}
\value{
tibble with lower-cased names
}
\description{
Load input by path, Excel, or data.frame
}
