% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-main.R
\name{er_main}
\alias{er_main}
\title{High-level facade (full pipeline + plotting/tables)}
\usage{
er_main(
  data,
  truth = NULL,
  fields = NULL,
  extra_fields = NULL,
  id_col = NULL,
  embed_col = NULL,
  k_clusters = 10,
  knn_k = 15,
  mst_cut_ratio = 5,
  mst_k = NULL,
  sn_window = 40,
  sn_method = "jw",
  sn_thresh = 0.12,
  svd_dim = 100,
  louvain_min_sim = 0,
  cos_thresh = 0.88,
  gc_thresholds = seq(0.06, 0.2, 0.02),
  gc_method = "rlf",
  gc_dist_method = "jw",
  auto_tune = TRUE,
  tune_metric = "adj_rand",
  k_grid = c(5, 10, 15, 20),
  eval_mode = "labeled_only",
  write_csv = NULL,
  sheet = NULL,
  run_comm_methods = c("walktrap", "infomap", "fast_greedy", "label_prop"),
  auto_plot = TRUE,
  show_tables = TRUE,
  show_progress = TRUE
)
}
\description{
High-level facade (full pipeline + plotting/tables)
}
