% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_measure_tm}
\alias{er_measure_tm}
\title{Measure time & memory of an expression}
\usage{
er_measure_tm(expr)
}
\value{
list(time_sec, peak_mem_MB, result)
}
\description{
Measure time & memory of an expression
}
\keyword{internal}
