% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-eval-truth.R
\name{er_pairs_to_clusters}
\alias{er_pairs_to_clusters}
\title{Convert pairwise truth to clusters}
\usage{
er_pairs_to_clusters(truth_pairs, id1 = "id1", id2 = "id2")
}
\description{
Convert pairwise truth to clusters
}
