% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_penalize_silhouette}
\alias{er_penalize_silhouette}
\title{Penalized Silhouette (avoid monotone-in-k)}
\usage{
er_penalize_silhouette(sil, k, n, alpha = NULL)
}
\description{
Penalized Silhouette (avoid monotone-in-k)
}
