% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30-reporting.R
\name{er_pick_gc_metric}
\alias{er_pick_gc_metric}
\title{Pick GC metric name from tuning table}
\usage{
er_pick_gc_metric(gc_curve, fallback_metric = "adj_rand")
}
\description{
Pick GC metric name from tuning table
}
\keyword{internal}
