% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-progress.R
\name{er_progress_done}
\alias{er_progress_done}
\title{Progress bar (done)}
\usage{
er_progress_done(p)
}
\arguments{
\item{p}{Progress object returned by \code{\link[=er_progress_start]{er_progress_start()}}.}
}
\value{
\code{NULL} invisibly. Closes the text progress bar and prints elapsed time.
}
\description{
Progress bar (done)
}
