% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-progress.R
\name{er_progress_start}
\alias{er_progress_start}
\title{Progress bar (start)}
\usage{
er_progress_start(total_steps, title = "ER pipeline")
}
\arguments{
\item{total_steps}{integer}

\item{title}{character}
}
\value{
progress environment
}
\description{
Progress bar (start)
}
