% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-progress.R
\name{er_progress_tick}
\alias{er_progress_tick}
\title{Progress bar (tick)}
\usage{
er_progress_tick(p, label = NULL)
}
\arguments{
\item{p}{progress object from \link{er_progress_start}}

\item{label}{optional label}
}
\description{
Progress bar (tick)
}
