% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-io.R
\name{er_read_pipe_or_fix}
\alias{er_read_pipe_or_fix}
\title{Read "pipe" delimited or try to fix common separators}
\usage{
er_read_pipe_or_fix(path)
}
\arguments{
\item{path}{file path}
}
\value{
data.frame with lower-cased names
}
\description{
Read "pipe" delimited or try to fix common separators
}
