% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-select.R
\name{er_select_fields}
\alias{er_select_fields}
\title{Select ID + build normalized text_for_matching column}
\usage{
er_select_fields(
  df,
  id_col = NULL,
  fields = NULL,
  extra_fields = NULL,
  id_candidates = c("id", "affiliation_id", "record_id", "rec_id", "docid", "rowid",
    "paper_id"),
  text_candidates = c("title", "name", "raw_title", "string", "text", "affiliation",
    "aggregate value", "clean ag.value"),
  embed_candidates = c("embedded clean ag.value", "embedded ag.value", "emb",
    "embedding", "vector", "embedding_clean"),
  normalize = TRUE,
  auto_fields = TRUE,
  max_auto_fields = 5
)
}
\arguments{
\item{df}{data.frame}

\item{id_col}{optional id name}

\item{fields}{vector of fields to concatenate}

\item{normalize}{logical}

\item{auto_fields}{logical}
}
\value{
tibble with id and text_for_matching
}
\description{
Select ID + build normalized text_for_matching column
}
