% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-distances-and-sil.R
\name{er_silhouette_avg}
\alias{er_silhouette_avg}
\title{Average silhouette from distance}
\usage{
er_silhouette_avg(labels, D)
}
\arguments{
\item{labels}{vector}

\item{D}{matrix or dist}
}
\value{
numeric
}
\description{
Average silhouette from distance
}
