% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_tfidf_svd}
\alias{er_tfidf_svd}
\title{Text2Vec TF-IDF + SVD embeddings}
\usage{
er_tfidf_svd(text, svd_dim = 200, ngram = c(1, 2), max_features = 1e+05)
}
\arguments{
\item{text}{character vector}

\item{svd_dim}{integer, default 200}

\item{ngram}{c(lo, hi), default c(1,2)}

\item{max_features}{max vocab size}
}
\value{
numeric matrix n x svd_dim
}
\description{
Text2Vec TF-IDF + SVD embeddings
}
