% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_threshold_cc}
\alias{er_threshold_cc}
\title{Thresholded connected components on cosine similarity}
\usage{
er_threshold_cc(Z, min_sim = 0.6, k_knn = 100)
}
\description{
Thresholded connected components on cosine similarity
}
