% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22-eval-truth.R
\name{er_truth_from_any}
\alias{er_truth_from_any}
\title{Truth ingestion from various formats}
\usage{
er_truth_from_any(
  truth,
  sep_pair = "\\\\|",
  id_candidates = c("id", "record_id", "rec_id", "docid", "rowid", "paper_id")
)
}
\description{
Truth ingestion from various formats
}
