% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{erbot_prep}
\alias{erbot_prep}
\title{Prepare Raw Data for Deduplication}
\usage{
erbot_prep(data, cols = NULL)
}
\arguments{
\item{data}{A data.frame or tibble of raw records}

\item{cols}{A character vector of the columns to standardize;
if NULL, all columns will be processed}
}
\value{
A tibble with cleaned, standardized fields
}
\description{
Prepare Raw Data for Deduplication
}
