% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/43-graph_utils.R
\name{graph_laplacian_matrix}
\alias{graph_laplacian_matrix}
\title{Graph Laplacian matrix}
\usage{
graph_laplacian_matrix(g)
}
\arguments{
\item{g}{An igraph graph with 'weight' on edges.}
}
\value{
A sparse Matrix object representing the Laplacian.
}
\description{
Compute the unnormalized Laplacian L = D - W
where W is the weighted adjacency matrix.
}
