% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/43-graph_utils.R
\name{graph_smoothness_loss}
\alias{graph_smoothness_loss}
\title{Graph smoothness loss}
\usage{
graph_smoothness_loss(Z, L)
}
\arguments{
\item{Z}{Numeric matrix \code{c([n, d])} of embeddings.}

\item{L}{Sparse Laplacian matrix \code{c([n, n])}.}
}
\value{
Numeric scalar (smoothness loss).
}
\description{
Compute Tr(Z^T L Z) as a scalar regularizer that encourages
embedding smoothness over the graph structure.
}
