% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-io.R
\name{ncvr_guess_fields}
\alias{ncvr_guess_fields}
\title{Guess NCVR-like fields}
\usage{
ncvr_guess_fields(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
character vector of fields
}
\description{
Guess NCVR-like fields
}
