% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-io.R
\name{ncvr_read}
\alias{ncvr_read}
\title{Read NC voter registration CSV splits}
\usage{
ncvr_read(root, which = c("5", "10", "all"))
}
\arguments{
\item{root}{folder}

\item{which}{"5","10","all"}
}
\value{
tibble
}
\description{
Read NC voter registration CSV splits
}
