% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-diagnose.R
\name{print.er_diagnose}
\alias{print.er_diagnose}
\title{Print an \code{er_diagnose} result}
\usage{
\method{print}{er_diagnose}(x, ...)
}
\arguments{
\item{x}{Named list returned by \code{\link{er_diagnose}()}.}

\item{...}{Currently unused; for S3 method consistency.}
}
\value{
\code{x} invisibly.
}
\description{
Print an \code{er_diagnose} result
}
