% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-ablation.R
\name{print_ablation}
\alias{print_ablation}
\title{Print an ablation result in a compact table}
\usage{
print_ablation(x, ...)
}
\arguments{
\item{x}{Output of \code{\link{er_ablation_all}}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print an ablation result in a compact table
}
