% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/43-graph_utils.R
\name{projection_from_membership}
\alias{projection_from_membership}
\title{Build projection matrix from cluster membership}
\usage{
projection_from_membership(membership)
}
\arguments{
\item{membership}{Integer membership vector, length n.}
}
\value{
Numeric matrix P with shape of \code{c([n, n])}.
}
\description{
Given a clustering, construct P = H H^T
where H has columns normalized by cluster size.
}
