% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/45-training_loop.R
\name{run_training}
\alias{run_training}
\title{Run training for multiple epochs}
\usage{
run_training(
  X_mat,
  pos_pairs,
  neg_pairs,
  epochs = 50,
  lr = 0.001,
  hidden_dim = 256,
  emb_dim = 64,
  k_nn = 10,
  lambda_graph = 0.001,
  lambda_K = 0.001
)
}
\arguments{
\item{X_mat}{Numeric matrix \code{c([n, p])} of input features.}

\item{pos_pairs}{Data frame of positive pairs (i, j).}

\item{neg_pairs}{Data frame of negative pairs (i, j).}

\item{epochs}{Number of training epochs.}

\item{lr}{Learning rate for Adam.}

\item{hidden_dim}{Hidden layer size for encoder.}

\item{emb_dim}{Embedding dimension.}

\item{k_nn}{Number of neighbours in kNN graph.}

\item{lambda_graph}{Weight of graph smoothness loss.}

\item{lambda_K}{Weight of stability loss.}
}
\value{
A list with elements:
\itemize{
\item encoder: trained RecordEncoder
\item history: data.frame of loss values by epoch
}
}
\description{
High-level driver that trains the RecordEncoder and
returns the trained encoder plus a history of loss values.
}
