% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{try_silent}
\alias{try_silent}
\title{Silent try}
\usage{
try_silent(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
The result of \code{expr}, or a \code{try-error} object.
}
\description{
Wraps \code{\link[base]{try}} with \code{silent = TRUE}.
}
\keyword{internal}
