% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_agglomerative_cosine}
\alias{er_agglomerative_cosine}
\title{Agglomerative (average linkage) on cosine distance}
\usage{
er_agglomerative_cosine(Z, k)
}
\description{
Agglomerative (average linkage) on cosine distance
}
