% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_cpm_score}
\alias{er_cpm_score}
\title{CPM quality function for a partition on a graph}
\usage{
er_cpm_score(g, membership, gamma = 0.05, weights = NULL)
}
\arguments{
\item{g}{igraph graph}

\item{membership}{integer labels}

\item{gamma}{numeric resolution parameter}

\item{weights}{optional numeric vector; defaults to E(g)$weight}
}
\description{
CPM quality function for a partition on a graph
}
\keyword{internal}
