% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_db_ch_indices}
\alias{er_db_ch_indices}
\title{Davies–Bouldin and Calinski–Harabasz indices}
\usage{
er_db_ch_indices(X, membership)
}
\arguments{
\item{X}{numeric matrix (rows = samples)}

\item{membership}{integer labels}
}
\description{
Davies–Bouldin and Calinski–Harabasz indices
}
\keyword{internal}
