% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_external_scores}
\alias{er_external_scores}
\title{External scores (pairwise P/R/F1, B³ P/R/F1, ARI)}
\usage{
er_external_scores(pred, truth)
}
\arguments{
\item{pred}{integer/factor vector of predicted clusters}

\item{truth}{integer/factor vector of gold clusters}
}
\description{
External scores (pairwise P/R/F1, B³ P/R/F1, ARI)
}
