% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-gc.R
\name{er_get_resolve_entities_fun}
\alias{er_get_resolve_entities_fun}
\title{Locate resolve_entities() (GCMER or sourced)}
\usage{
er_get_resolve_entities_fun()
}
\value{
function
}
\description{
Locate resolve_entities() (GCMER or sourced)
}
\keyword{internal}
