% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-methods-basic.R
\name{er_hclust_from_X}
\alias{er_hclust_from_X}
\title{Hierarchical clustering (Ward.D2) with cosine distance}
\usage{
er_hclust_from_X(X, k = 10, method = "ward.D2")
}
\description{
Hierarchical clustering (Ward.D2) with cosine distance
}
