% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_leiden_from_Z}
\alias{er_leiden_from_Z}
\title{Leiden on kNN graph (via leidenbase)}
\usage{
er_leiden_from_Z(
  Z,
  k_knn = 50,
  min_sim = 0.1,
  resolution_parameter = 0.5,
  partition_type = "ModularityVertexPartition"
)
}
\arguments{
\item{resolution_parameter}{numeric; higher -> more clusters}

\item{partition_type}{"ModularityVertexPartition" or "CPMVertexPartition"}
}
\description{
Leiden on kNN graph (via leidenbase)
}
