% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/32-main.R
\name{er_main_simple}
\alias{er_main_simple}
\title{Minimal facade (quick comparisons)}
\usage{
er_main_simple(
  data,
  truth = NULL,
  fields = c("title", "authors", "address"),
  k_grid = seq(10, 300, by = 10),
  write_csv = NULL,
  svd_k = 100,
  knn_k = 15,
  louvain_min_sim = 0,
  use_methods = c("kmeans_tfidf_svd", "louvain_knn", "louvain_multifield")
)
}
\description{
Minimal facade (quick comparisons)
}
