% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_mean_ari}
\alias{er_mean_ari}
\title{Mean ARI across a list of label vectors (stability)}
\usage{
er_mean_ari(labels_list)
}
\arguments{
\item{labels_list}{list of integer vectors (same length)}
}
\description{
Mean ARI across a list of label vectors (stability)
}
