% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_measure_tm}
\alias{er_measure_tm}
\title{Measure time & memory of an expression}
\usage{
er_measure_tm(expr)
}
\arguments{
\item{expr}{An R expression to evaluate.}
}
\value{
Named list with \code{time_sec} (elapsed seconds), \code{peak_mem_MB}
(peak RAM in MiB), and \code{result} (the value of \code{expr}).
}
\description{
Measure time & memory of an expression
}
\keyword{internal}
