% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-evaluate.R
\name{er_pairs_to_clusters}
\alias{er_pairs_to_clusters}
\title{Convert pairwise match list to cluster labels}
\usage{
er_pairs_to_clusters(truth_pairs, id1 = "id1", id2 = "id2")
}
\arguments{
\item{truth_pairs}{data.frame with columns \code{id1}, \code{id2}.}
}
\value{
tibble(id, cluster_id).
}
\description{
Convert pairwise match list to cluster labels
}
