% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-distances-and-sil.R
\name{er_pairwise_stringdist}
\alias{er_pairwise_stringdist}
\title{Pairwise string distances in blocks (with optional progress)}
\usage{
er_pairwise_stringdist(text_vec, method = "jw", block = 4000, progress = NULL)
}
\description{
Pairwise string distances in blocks (with optional progress)
}
