% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-reporting.R
\name{er_plot_curve}
\alias{er_plot_curve}
\title{Quick base plotting for a tuning curve}
\usage{
er_plot_curve(df, x_col, y_col, main = "", xlab = NULL, ylab = NULL)
}
\description{
Quick base plotting for a tuning curve
}
