% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-progress.R
\name{er_progress_tick}
\alias{er_progress_tick}
\title{Progress bar (tick)}
\usage{
er_progress_tick(p, label = NULL)
}
\arguments{
\item{p}{Progress object returned by \code{\link[=er_progress_start]{er_progress_start()}}.}

\item{label}{Optional character label printed beside the step counter.}
}
\value{
The updated progress object (invisibly).
}
\description{
Progress bar (tick)
}
