% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-reporting.R
\name{er_save_report_pdf}
\alias{er_save_report_pdf}
\title{Save an organized ER PDF report (backward-compatible, no Rmd)}
\usage{
er_save_report_pdf(
  res,
  file,
  dataset_name = "DATASET",
  top_n = 5,
  digits = 5,
  runtime_sec = NULL
)
}
\arguments{
\item{res}{ER result from er_main() or similar. Handles old/new field layouts.}

\item{file}{Output PDF path.}

\item{dataset_name}{Title label.}

\item{top_n}{Top items to show if res$top_items present.}

\item{digits}{Rounding for numbers.}

\item{runtime_sec}{Optional total runtime in seconds.}
}
\description{
Save an organized ER PDF report (backward-compatible, no Rmd)
}
