% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-reporting.R
\name{er_save_report_pdf}
\alias{er_save_report_pdf}
\title{Save a PDF report for a pipeline result}
\usage{
er_save_report_pdf(
  res,
  file = "er_report.pdf",
  dataset_name = NULL,
  top_n = 5,
  metric = NULL,
  width = 11,
  height = 8.5
)
}
\arguments{
\item{res}{result list returned by \link{er_unified_pipeline}}

\item{file}{output pdf path}

\item{dataset_name}{character}
}
\description{
Save a PDF report for a pipeline result
}
