% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50-tuning.R
\name{er_scaling_curve}
\alias{er_scaling_curve}
\title{Scalability curves by subsampling n}
\usage{
er_scaling_curve(
  data,
  fields,
  n_seq,
  method = "kmeans",
  params_fixed = list(),
  svd_dim = 200
)
}
\description{
Scalability curves by subsampling n
}
