% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/33-weighted-pipeline.R
\name{expand_fields_by_weights}
\alias{expand_fields_by_weights}
\title{Expand fields proportionally to weights (fallback)}
\usage{
expand_fields_by_weights(fields, weights, base_rep = 10L)
}
\description{
Expand fields proportionally to weights (fallback)
}
