% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R, R/50-tuning.R
\name{\%||\%}
\alias{\%||\%}
\title{Internal utility: Null-coalescing}
\usage{
a \%||\% b

a \%||\% b
}
\description{
Turnkey utilities to (1) tune hyperparameters with robust internal criteria,
(2) run mainstream ER methods (kmeans/agglo/DBSCAN/Louvain/Leiden/CW/threshold-CC/MST+edit),
and (3) measure time & memory and generate scalability curves.
}
\keyword{benchmarking}
\keyword{clustering,}
\keyword{entity-resolution,}
\keyword{graph,}
\keyword{internal}
\keyword{tuning,}
