% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{ncvr_guess_fields}
\alias{ncvr_guess_fields}
\title{Guess NCVR-like fields}
\usage{
ncvr_guess_fields(df)
}
\arguments{
\item{df}{data.frame whose column names will be searched.}
}
\value{
Character vector of matched column names, or \code{NULL}.
}
\description{
Guess NCVR-like fields
}
