% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-load.R
\name{ncvr_read}
\alias{ncvr_read}
\title{Read NC voter registration CSV splits}
\usage{
ncvr_read(root, which = c("5", "10", "all"))
}
\arguments{
\item{root}{Directory containing CSV partitions.}

\item{which}{Partition set: \code{"5"}, \code{"10"}, or \code{"all"}.}
}
\value{
A \code{tibble} with lower-cased column names.
}
\description{
Read NC voter registration CSV splits
}
